Stop the leak, use impredicativity!
Upgrade the component to use my new impredicativity trick to hide the query, while still allowing arbitrary output to be passed through (of kind Type). Also upgrade to allow it to be a transparent parent component, providing an option that defaults to not being a parent component to avoid ambiguous type variables.